refactor(core): isolate harness integration paths
