+++ integrations/oh-my-pi/connector/src/agent-dir.ts
@@ -1,7 +1,9 @@
 import { existsSync, mkdirSync, readFileSync, rmSync, writeFileSync } from "node:fs";
 import { homedir } from "node:os";
 import { dirname, join, resolve } from "node:path";
-import { expandHome } from "./constants.js";
+
+const OH_MY_PI_MANAGED_FILENAMES = ["signet-oh-my-pi.js", "signet-oh-my-pi.mjs"] as const;
+const OH_MY_PI_MANAGED_MARKER = "SIGNET_MANAGED_OH_MY_PI_EXTENSION";
 
 interface OhMyPiConfigFile {
 	readonly version: 1;
@@ -16,13 +18,25 @@ function readTrimmed(env: NodeJS.ProcessEnv, name: string): string | null {
 	return trimmed.length > 0 ? trimmed : null;
 }
 
-function normalizePath(pathValue: string): string {
-	return resolve(expandHome(pathValue.trim()));
+function userHome(env: NodeJS.ProcessEnv): string {
+	return env.HOME?.trim() || homedir();
+}
+
+function expandUserPath(pathValue: string, env: NodeJS.ProcessEnv): string {
+	const trimmed = pathValue.trim();
+	const home = userHome(env);
+	if (trimmed === "~") return home;
+	if (trimmed.startsWith("~/") || trimmed.startsWith("~\\")) return join(home, trimmed.slice(2));
+	return trimmed;
+}
+
+function normalizePath(pathValue: string, env: NodeJS.ProcessEnv): string {
+	return resolve(expandUserPath(pathValue, env));
 }
 
 function readConfigHome(env: NodeJS.ProcessEnv): string {
 	const configured = readTrimmed(env, "XDG_CONFIG_HOME");
-	return configured ? normalizePath(configured) : join(homedir(), ".config");
+	return configured ? normalizePath(configured, env) : join(userHome(env), ".config");
 }
 
 export function getOhMyPiConfigPath(env: NodeJS.ProcessEnv = process.env): string {
@@ -37,7 +51,7 @@ export function readConfiguredOhMyPiAgentDir(env: NodeJS.ProcessEnv = process.en
 		const raw: unknown = JSON.parse(readFileSync(configPath, "utf-8"));
 		if (typeof raw !== "object" || raw === null) return null;
 		const agentDir = Reflect.get(raw, "agentDir");
-		return typeof agentDir === "string" && agentDir.trim().length > 0 ? normalizePath(agentDir) : null;
+		return typeof agentDir === "string" && agentDir.trim().length > 0 ? normalizePath(agentDir, env) : null;
 	} catch {
 		return null;
 	}
@@ -52,8 +66,8 @@ export function readConfiguredOhMyPiAgentDir(env: NodeJS.ProcessEnv = process.en
  */
 export function resolveOhMyPiAgentDir(env: NodeJS.ProcessEnv = process.env): string {
 	const configured = readTrimmed(env, "PI_CODING_AGENT_DIR");
-	if (configured) return normalizePath(configured);
-	return readConfiguredOhMyPiAgentDir(env) ?? join(homedir(), ".omp", "agent");
+	if (configured) return normalizePath(configured, env);
+	return readConfiguredOhMyPiAgentDir(env) ?? join(userHome(env), ".omp", "agent");
 }
 
 export function resolveOhMyPiExtensionsDir(env: NodeJS.ProcessEnv = process.env): string {
@@ -63,15 +77,32 @@ export function resolveOhMyPiExtensionsDir(env: NodeJS.ProcessEnv = process.env)
 export function listOhMyPiAgentDirCandidates(env: NodeJS.ProcessEnv = process.env): readonly string[] {
 	const candidates = new Set<string>();
 	const configured = readTrimmed(env, "PI_CODING_AGENT_DIR");
-	if (configured) candidates.add(normalizePath(configured));
+	if (configured) candidates.add(normalizePath(configured, env));
 	const persisted = readConfiguredOhMyPiAgentDir(env);
 	if (persisted) candidates.add(persisted);
-	candidates.add(join(homedir(), ".omp", "agent"));
+	candidates.add(join(userHome(env), ".omp", "agent"));
 	return Array.from(candidates);
 }
 
+export function hasOhMyPiSetup(env: NodeJS.ProcessEnv = process.env): boolean {
+	if (existsSync(resolveOhMyPiAgentDir(env))) return true;
+	for (const agentDir of listOhMyPiAgentDirCandidates(env)) {
+		const extensionsDir = join(agentDir, "extensions");
+		for (const filename of OH_MY_PI_MANAGED_FILENAMES) {
+			const extensionPath = join(extensionsDir, filename);
+			if (!existsSync(extensionPath)) continue;
+			try {
+				if (readFileSync(extensionPath, "utf8").includes(OH_MY_PI_MANAGED_MARKER)) return true;
+			} catch {
+				// Ignore unreadable candidate and continue checking others.
+			}
+		}
+	}
+	return false;
+}
+
 export function writeConfiguredOhMyPiAgentDir(pathValue: string, env: NodeJS.ProcessEnv = process.env): string {
-	const agentDir = normalizePath(pathValue);
+	const agentDir = normalizePath(pathValue, env);
 	const configPath = getOhMyPiConfigPath(env);
 	if (readConfiguredOhMyPiAgentDir(env) === agentDir && existsSync(configPath)) {
 		return configPath;
