+++ integrations/pi/connector/package.json
@@ -26,7 +26,7 @@
 	},
 	"dependencies": {
 		"@signet/connector-base": "0.219.6",
-		"@signet/core": "0.219.6"
+		"@signet/pi-extension-base": "workspace:*"
 	},
 	"devDependencies": {
 		"@types/node": "^22.0.0",
