+++ integrations/pi/extension-base/package.json
@@ -6,7 +6,8 @@
 	"main": "./src/index.ts",
 	"types": "./src/index.ts",
 	"exports": {
-		".": "./src/index.ts"
+		".": "./src/index.ts",
+		"./agent-dir": "./src/agent-dir.ts"
 	},
 	"dependencies": {
 		"@signet/core": "workspace:*",
