+++ surfaces/cli/src/features/setup-embedding-validation.test.ts
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, it, mock } from "bun:test";
 import { mkdtempSync, readFileSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import type { SetupDetection } from "@signet/core";
+import type { SetupDetection } from "../lib/setup-detection.js";
 import { validateOllamaModelNonInteractive } from "./setup-providers.js";
 import type { SetupDeps } from "./setup-types.js";
 import { setupWizard } from "./setup.js";
